Adding my demo files
Here are the files for the demo of the game engine that i have produced.
It should be run with on windows with g++ (look into mingw), and can be
run on linux if the .a files and the .dll files are swapped for their
appropriate counterparts. See
http://codeumbra.eu/complete-roguelike-tutorial-using-c-and-libtcod-part-1-setting-up

for more info on creating a proper runtime environment.