Fixed some infection-related bugs
Made the map::save and map::load functions save and load the tile
infection levels. Made the map only place an epicenter and infected
tiles when a new game is started, not also on game loading.

Got rid of actor->enviroment, as it was a waste for what it was being
used for. We had access to map.tiles anyway, and saving a bunch of
copies of the map wasn't worth it.