Got rid of a bug, kept another
Got rid of the bug where when you save in the first room or so, the game
makes a ton of floor tiles into walls. No idea why at the moment.

This introduces a bug where when you load, a new epicenter is made every
time, so the screen is flooded a bit every 500 turns when all of them
update at once.