Added vending machines
I also added a system for the player to be neutral towards actors that
are also neutral, which is his default action. He can choose to be
hostile towards all actors by pressing '`'. The tilde key. If the player
is in a neutral stance, and the monster he bumps into has an interaction
it can perform with the player, he does so. This is achieved via the
InteractibleAi Ai type, of which the vending Ai is a subtype. Right now,
vending machines softly hum, but the vendor Gui stuff can be put there
instead.