Aura class implemented, hunger added
Feel free to tweak the numbers on hunger. I think next there should be a
function to recalculate stat values after debuffs have been applied. An
actor function, I think, to recalculate stats and all that.

The aura class may not be the best solution or even perfect code, but it
is there. I can explain whatever needs explaining.