Container now handles inventory fullness better
Now the inventory should manage checking whether it is full better.
Maybe not, though.

The issue prior to this commit is that after we added paging and
stacking for the inventory, we never changed over to a system them that
limits each sort type to a size of 26 UNIQUE items. I attempted to do
that with this commit.